extract WatchCharacteristic component from WatchCharacteristics
